build: track symbols
for easier auto-generation of versioned deps. when adding new symbols,
the build should display a warning + diff (in addition to our manual
tracking of the generated header file). changes in symbol signatures or
semantics are not caught automatically by the symbols file, but still
require bumping the minimum library version for that symbol accordingly.

Signed-off-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>