fix: never append archive name extensions twice, use type instead
Commit 16298c5f ("update to current PBS master") introduced a helper
method to generate the backup archive names from a device name.

However, the helper was also used incorrectly to generate archive
names from strings already having the `.img.fidx` filename extension.
As the extension is appended unconditionally, this leads to an
archive name which will parse successfully, but not be the expected
archive name to be found in the backup manifest.

Fix this by dropping the incorrect helper call-sites and adapt the
function parameters to take a `BackupArchiveName` directly, reducing
potential for future mishandling.

Fixes: Commit 16298c5f ("update to current PBS master")
Reported-by: Markus Frank <m.frank@proxmox.com>
Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Tested-by: Markus Frank <m.frank@proxmox.com>